refactor: promote shared modules to app/libs and app/components
Move widely-referenced modules out of route directories into proper
shared locations:

- classify.ts (11 consumers) → app/libs/pr-classify.ts
  Import: ~/app/libs/pr-classify

- size-badge.tsx (6 consumers) → app/components/size-badge.tsx
  Import: ~/app/components/size-badge

- size-badge-popover.tsx (5 consumers) → app/components/size-badge-popover.tsx
  Import: ~/app/components/size-badge-popover

These modules were in route-scoped +functions/+components directories
but were referenced across workload, throughput, analysis, and other
routes. Promoting them eliminates deep relative/absolute paths like
~/app/routes/$orgSlug/analysis/reviews/+functions/classify.

Closes #184
